% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxGrid.R
\name{approx.grid}
\alias{approx.grid}
\title{Linear interpolate between 2 parallel grids}
\usage{
approx.grid(x, y, xout)
}
\arguments{
\item{x}{vector of 2 scalars}

\item{y}{list of 2 grids (must be equi-spaced)}

\item{xout}{scalar to interpolate at}
}
\value{
A grid of interpolated values (or NA values), with dimension matching that of either grid in y.
}
\description{
Linear interpolate between 2 parallel grids
}
