% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate2dGrid.R
\name{covariate.2d.grid}
\alias{covariate.2d.grid}
\title{Calculate covariates at grid locations and pred.time,
using trilinear (2D + T) interpolation}
\usage{
covariate.2d.grid(pred.grid, COV, pred.time)
}
\arguments{
\item{COV}{MLD, RAD, WIN}
}
\value{
Covariates at grid locations and pred.time.
}
\description{
Calculate covariates at grid locations and pred.time,
using trilinear (2D + T) interpolation
}
