% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate2d.R
\name{covariate.2d}
\alias{covariate.2d}
\title{Calculate covariates at gldr.mean locations, using bilinear (2D) interpolation}
\usage{
covariate.2d(
  train.length,
  training,
  gldr.mean,
  cov.lon,
  cov.lat,
  cov.training,
  cov.var
)
}
\arguments{
\item{cov}{rad}
}
\value{
Covariates at gldr.mean locations.
}
\description{
Calculate covariates at gldr.mean locations, using bilinear (2D) interpolation
}
