% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate2dtGrid.R
\name{covariate.2dt.grid}
\alias{covariate.2dt.grid}
\title{Calculate covariates at grid locations, using trilinear (2D + T) interpolation}
\usage{
covariate.2dt.grid(
  grid.list,
  dat.lon,
  dat.lat,
  dat.time,
  dat.time.predicting,
  k,
  variable
)
}
\arguments{
\item{dat.lon}{mld.lon, ...}

\item{variable}{mlotst, ...}
}
\value{
Covariates at grid locations.
}
\description{
Calculate covariates at grid locations, using trilinear (2D + T) interpolation
}
