% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate2dt.R
\name{covariate.2dt}
\alias{covariate.2dt}
\title{Calculate covariates at gldr.mean locations, using trilinear (2D + T) interpolation}
\usage{
covariate.2dt(
  train.length,
  training,
  gldr.mean,
  cov.lon,
  cov.lat,
  cov.time,
  cov.time.training,
  cov.var
)
}
\arguments{
\item{cov}{mld}
}
\value{
Covariates at gldr.mean locations.
}
\description{
Calculate covariates at gldr.mean locations, using trilinear (2D + T) interpolation
}
