% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3dGrid.R
\name{covariate.3d.grid}
\alias{covariate.3d.grid}
\title{Calculate covariates at grid locations and pred.time,
using quadrilinear (3D + T) interpolation}
\usage{
covariate.3d.grid(pred.grid, pred.depth, COV, pred.time)
}
\arguments{
\item{COV}{CHL, CUX, CUY, OXY, NIT, PHO, TEM, TUR}
}
\value{
Covariates at grid locations and pred.time.
}
\description{
Calculate covariates at grid locations and pred.time,
using quadrilinear (3D + T) interpolation
}
