% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3d.R
\name{covariate.3d}
\alias{covariate.3d}
\title{Calculate covariates at gldr.mean locations, using trilinear (3D) interpolation}
\usage{
covariate.3d(
  train.length,
  training,
  gldr.mean,
  cov.lon,
  cov.lat,
  cov.depth,
  cov.training,
  cov.var
)
}
\arguments{
\item{cov}{tem, cur}
}
\value{
Covariates at gldr.mean locations.
}
\description{
Calculate covariates at gldr.mean locations, using trilinear (3D) interpolation
}
