% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3dtAltGrid.R
\name{covariate.3dt.alt.grid}
\alias{covariate.3dt.alt.grid}
\title{Calculate covariates at grid locations, using quadrilinear (3D + T) interpolation
in an alternative order (depth, time, lon, lat)}
\usage{
covariate.3dt.alt.grid(
  grid.list,
  gldr.depth.mean,
  cov.lon,
  cov.lat,
  cov.depth,
  cov.time,
  cov.time.predicting,
  k,
  cov.var
)
}
\arguments{
\item{cov}{ass.chl}
}
\value{
Covariates at grid locations.
}
\description{
Calculate covariates at grid locations, using quadrilinear (3D + T) interpolation
in an alternative order (depth, time, lon, lat)
}
