% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3dtAlt.R
\name{covariate.3dt.alt}
\alias{covariate.3dt.alt}
\title{Calculate covariates at gldr.mean locations, using quadrilinear (3D + T) interpolation
in an alternative order (depth, time, lon, lat)}
\usage{
covariate.3dt.alt(
  train.length,
  training,
  gldr.mean,
  cov.lon,
  cov.lat,
  cov.depth,
  cov.time,
  cov.time.training,
  cov.var
)
}
\arguments{
\item{cov}{ass.chl}
}
\value{
Covariates at gldr.mean locations.
}
\description{
Calculate covariates at gldr.mean locations, using quadrilinear (3D + T) interpolation
in an alternative order (depth, time, lon, lat)
}
