% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3dtGrid.R
\name{covariate.3dt.grid}
\alias{covariate.3dt.grid}
\title{Calculate covariates at grid locations, using quadrilinear (3D + T) interpolation}
\usage{
covariate.3dt.grid(
  grid.list,
  gldr.depth.mean,
  cov.lon,
  cov.lat,
  cov.depth,
  cov.time,
  cov.time.predicting,
  k,
  cov.var
)
}
\arguments{
\item{cov}{chl, oxy, nit, pho, tur}
}
\value{
Covariates at grid locations.
}
\description{
Calculate covariates at grid locations, using quadrilinear (3D + T) interpolation
}
