% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3dt.R
\name{covariate.3dt}
\alias{covariate.3dt}
\title{Calculate covariates at gldr.mean locations, using quadrilinear (3D + T) interpolation}
\usage{
covariate.3dt(
  train.length,
  training,
  gldr.mean,
  dat.lon,
  dat.lat,
  dat.depth,
  dat.time,
  dat.time.training,
  variable
)
}
\arguments{
\item{dat.lon}{chl.lon, ...}

\item{variable}{chlor, ...}
}
\value{
Covariates at gldr.mean locations.
}
\description{
Calculate covariates at gldr.mean locations, using quadrilinear (3D + T) interpolation
}
