% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate4dGrid.R
\name{covariate.4d.grid}
\alias{covariate.4d.grid}
\title{Calculate covariates at grid locations and pred.time,
using quadrilinear (3D + T) interpolation
for a sigma coordinate system, where depth = f(lon, lat, depth, time)}
\usage{
covariate.4d.grid(pred.grid, pred.depth, COV, pred.time)
}
\arguments{
\item{COV}{MASS-4D CHL}
}
\value{
Covariates at grid locations and pred.time.
}
\description{
Calculate covariates at grid locations and pred.time,
using quadrilinear (3D + T) interpolation
for a sigma coordinate system, where depth = f(lon, lat, depth, time)
}
