% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateAlt3dGrid.R
\name{covariate.alt.3d.grid}
\alias{covariate.alt.3d.grid}
\title{Calculate covariates at grid locations and pred.time,
using quadrilinear (3D + T) interpolation
in an alternative order (depth, time, lon, lat)}
\usage{
covariate.alt.3d.grid(pred.grid, pred.depth, COV, pred.time)
}
\arguments{
\item{COV}{CHLF, CHLR, OXYF, OXYR}
}
\value{
Covariates at grid locations and pred.time.
}
\description{
Calculate covariates at grid locations and pred.time,
using quadrilinear (3D + T) interpolation
in an alternative order (depth, time, lon, lat)
}
