% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateAlt3d.R
\name{covariate.alt.3d}
\alias{covariate.alt.3d}
\title{Calculate covariates at train.data locations and train.times,
using quadrilinear (3D + T) interpolation
in an alternative order (depth, time, lon, lat)}
\usage{
covariate.alt.3d(train.length, train.data, COV, train.times)
}
\arguments{
\item{COV}{CHLF, CHLR, OXYF, OXYR}
}
\value{
Covariates at train.data locations and train.times.
}
\description{
Calculate covariates at train.data locations and train.times,
using quadrilinear (3D + T) interpolation
in an alternative order (depth, time, lon, lat)
}
