% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDepthGrid.R
\name{find.depth.grid}
\alias{find.depth.grid}
\title{Find nearest and neighbour depth indices when predicting}
\usage{
find.depth.grid(gldr.depth.mean, dat.depth)
}
\arguments{
\item{dat.depth}{tem.depth, cur.depth, chl.depth ...}
}
\value{
Nearest and neighbour depth indices.
}
\description{
Find nearest and neighbour depth indices when predicting
}
