% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDepth.R
\name{find.depth}
\alias{find.depth}
\title{Find nearest and neighbour depth indices}
\usage{
find.depth(gldr.mean, i, cov.depth)
}
\arguments{
\item{cov.depth}{tem.depth, cur.depth, chl.depth ...}
}
\value{
Nearest and neighbour depth indices when training
}
\description{
Find nearest and neighbour depth indices
}
