% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIndex.R
\name{find.index}
\alias{find.index}
\title{Find nearest and neighbour indices}
\usage{
find.index(arbitrary.values, i, standard.values)
}
\arguments{
\item{standard.values}{}
}
\value{
Nearest and neighbour indices.
}
\description{
Find nearest and neighbour indices
}
