% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTime.R
\name{find.time}
\alias{find.time}
\title{Find nearest and neighbour time indices}
\usage{
find.time(cov.time.simulating, i, cov.time)
}
\arguments{
\item{cov.time.simulating}{cov.time.training, cov.time.predicting}

\item{cov.time}{mld.time, chl.time, ...}
}
\value{
Nearest and neighbour time indices.
}
\description{
Find nearest and neighbour time indices
}
