% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaBounds.R
\name{inla.bounds}
\alias{inla.bounds}
\title{Identify whether grid.mean is out of bounds.}
\usage{
inla.bounds(
  grid.mean,
  l = 1,
  k,
  RData,
  lower.bound,
  upper.bound,
  train.data,
  count.limit
)
}
\arguments{
\item{grid.mean}{mean predictions on grid}
}
\value{
0 = in bounds; 1 = out of bounds.
}
\description{
Identify whether grid.mean is out of bounds.
}
