% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaError.R
\name{inla.error}
\alias{inla.error}
\title{Identify whether inla crashed.}
\usage{
inla.error(mod.mode, l = 1, k, RData)
}
\arguments{
\item{mod.mode}{result of inla}
}
\value{
0 = no crash; 1 = crash.
}
\description{
Identify whether inla crashed.
}
