% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaLog.R
\name{inla.log}
\alias{inla.log}
\title{Extract and save run time, stupid searches and warnings from log.
Returns number of warnings.}
\usage{
inla.log(mod.mode, l = 1, k, RData, directory, write = FALSE)
}
\arguments{
\item{mod.mode}{result of inla}

\item{directory}{output directory}

\item{write}{write the log file (TRUE or FALSE)}
}
\value{
Fixed effects mean and sd; range 0.025 and stdev 0.975 quantiles.
}
\description{
Extract and save run time, stupid searches and warnings from log.
Returns number of warnings.
}
