% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaWarning.R
\name{inla.warning}
\alias{inla.warning}
\title{Returns number of warnings from log.
If no warnings, extracts and saves run time and number of stupid searches from log.
If warning(s), prints "INLA warning(s); repeating calculation...".}
\usage{
inla.warning(mod.mode, l = 1, k, RData, directory, write = FALSE)
}
\arguments{
\item{mod.mode}{result of inla}

\item{directory}{output directory}

\item{write}{write the log file (TRUE or FALSE)}
}
\value{
0 = no warnings; !0 = warnings.
}
\description{
Returns number of warnings from log.
If no warnings, extracts and saves run time and number of stupid searches from log.
If warning(s), prints "INLA warning(s); repeating calculation...".
}
