% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData2d.R
\name{load.data.2d}
\alias{load.data.2d}
\title{Load covariate data, including 2D, time and 1 variable
Includes flag for data source}
\usage{
load.data.2d(file, source = "CMEMS")
}
\arguments{
\item{file}{path to file}

\item{source}{ECMWF}
}
\value{
Covariate data.
}
\description{
Load covariate data, including 2D, time and 1 variable
Includes flag for data source
}
