% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData3d.R
\name{load.data.3d}
\alias{load.data.3d}
\title{Load covariate data, including 3D, time and 1 variable.
Includes flags for variable and data source.
Also returns origin from which time is counted (excludes MASS data).}
\usage{
load.data.3d(file, variable = "thetao", source = "CMEMS")
}
\arguments{
\item{file}{path to file}

\item{variable}{uo, vo, chlf, chlr, oxyf, oxyr}

\item{source}{MASS}
}
\value{
Covariate data and origin.
}
\description{
Load covariate data, including 3D, time and 1 variable.
Includes flags for variable and data source.
Also returns origin from which time is counted (excludes MASS data).
}
