% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData3d1.R
\name{load.data.3d1}
\alias{load.data.3d1}
\title{Load covariate data, including 3D, time and 1 variable}
\usage{
load.data.3d1(data.file)
}
\arguments{
\item{data.file}{path to data file}
}
\value{
Covariate data.
}
\description{
Load covariate data, including 3D, time and 1 variable
}
