% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData3d2.R
\name{load.data.3d2}
\alias{load.data.3d2}
\title{Load covariate data, including 3D, time and 2 variables}
\usage{
load.data.3d2(data.file)
}
\arguments{
\item{data.file}{path to data file}
}
\value{
Covariate data.
}
\description{
Load covariate data, including 3D, time and 2 variables
}
