% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTimeIndex.R
\name{set.time.index}
\alias{set.time.index}
\title{Calculate training and predicting time indices}
\usage{
set.time.index(
  gldr.time.start,
  data.time.start,
  origin.time,
  cov.time,
  train.length,
  nt
)
}
\arguments{
\item{data.time.start}{time CMEMS & ECMWF records begin}

\item{origin.time}{time glider records begin}

\item{cov.time}{covariate time values}
}
\value{
Training and predicting time indices.
}
\description{
Calculate training and predicting time indices
}
