% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTime.R
\name{set.time}
\alias{set.time}
\title{Calculate training and predicting times}
\usage{
set.time(
  gldr.time.start,
  data.time.start,
  origin.time,
  train.length,
  nt,
  train.interval = 3600,
  pred.interval = 3600,
  flight.interval = 3600
)
}
\arguments{
\item{data.time.start}{time CMEMS & ECMWF records begin}

\item{origin.time}{time glider records begin}

\item{train.interval}{interval between training times}

\item{pred.interval}{interval between prediction times}

\item{flight.interval}{interval between training flight and prediction flight}
}
\value{
Training and predicting times.
}
\description{
Calculate training and predicting times
}
