% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTurbidityGrid.R
\name{sum.turbidity.grid}
\alias{sum.turbidity.grid}
\title{Sum turbidity to pred.depth at pred.grid locations and pred.time}
\usage{
\method{sum}{turbidity.grid}(pred.grid, pred.depth, TUR, pred.time)
}
\value{
Turbidity sum at pred.grid locations and pred.time.
}
\description{
Sum turbidity to pred.depth at pred.grid locations and pred.time
}
