% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTurbidity.R
\name{sum.turbidity}
\alias{sum.turbidity}
\title{Sum turbidity to train.data depths at train.data locations and train.times}
\usage{
\method{sum}{turbidity}(train.length, train.data, TUR, train.times)
}
\value{
Turbidity sum at train.data locations and train.times.
}
\description{
Sum turbidity to train.data depths at train.data locations and train.times
}
