% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate2d.R
\name{covariate.2d}
\alias{covariate.2d}
\title{Calculate covariates at train.data locations and train.times,
using trilinear (2D + T) interpolation}
\usage{
covariate.2d(train.length, train.data, COV, train.times)
}
\arguments{
\item{COV}{MLD, RAD, WIN}
}
\value{
Covariates at train.data locations and train.times.
}
\description{
Calculate covariates at train.data locations and train.times,
using trilinear (2D + T) interpolation
}
