% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate2dtGrid.R
\name{covariate.2dt.grid}
\alias{covariate.2dt.grid}
\title{Calculate covariates at grid locations, using trilinear (2D + T) interpolation}
\usage{
covariate.2dt.grid(
  grid.list,
  cov.lon,
  cov.lat,
  cov.time,
  cov.time.predicting,
  k,
  cov.var
)
}
\arguments{
\item{cov}{mld, rad, win}
}
\value{
Covariates at grid locations.
}
\description{
Calculate covariates at grid locations, using trilinear (2D + T) interpolation
}
