% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3dGrid.R
\name{covariate.3d.grid}
\alias{covariate.3d.grid}
\title{Calculate covariates at grid locations, using trilinear (3D) interpolation}
\usage{
covariate.3d.grid(
  grid.list,
  gldr.depth.mean,
  dat.lon,
  dat.lat,
  dat.depth,
  dat.predicting,
  k,
  variable
)
}
\arguments{
\item{dat.lon}{tem.lon, cur.lon, ...}

\item{variable}{thetao, uo, vo, ...}
}
\value{
Covariates at grid locations.
}
\description{
Calculate covariates at grid locations, using trilinear (3D) interpolation
}
