% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3d.R
\name{covariate.3d}
\alias{covariate.3d}
\title{Calculate covariates at train.data locations and train.times,
using quadrilinear (3D + T) interpolation}
\usage{
covariate.3d(train.length, train.data, COV, train.times)
}
\arguments{
\item{COV}{CHL, CUX, CUY, OXY, NIT, PHO, TEM, TUR}
}
\value{
Covariates at train.data locations and train.times.
}
\description{
Calculate covariates at train.data locations and train.times,
using quadrilinear (3D + T) interpolation
}
