% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3dtGrid.R
\name{covariate.3dt.grid}
\alias{covariate.3dt.grid}
\title{Calculate covariates at grid locations, using quadrilinear (3D + T) interpolation}
\usage{
covariate.3dt.grid(
  grid.list,
  gldr.depth.mean,
  dat.lon,
  dat.lat,
  dat.depth,
  dat.time,
  dat.time.predicting,
  k,
  variable
)
}
\arguments{
\item{dat.lon}{chl.lon, ...}

\item{variable}{chlor, ...}
}
\value{
Covariates at grid locations.
}
\description{
Calculate covariates at grid locations, using quadrilinear (3D + T) interpolation
}
