% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate3dt.R
\name{covariate.3dt}
\alias{covariate.3dt}
\title{Calculate covariates at gldr.mean locations, using quadrilinear (3D + T) interpolation}
\usage{
covariate.3dt(
  train.length,
  training,
  gldr.mean,
  cov.lon,
  cov.lat,
  cov.depth,
  cov.time,
  cov.time.training,
  cov.var
)
}
\arguments{
\item{cov}{chl, oxy, nit, pho, tur}
}
\value{
Covariates at gldr.mean locations.
}
\description{
Calculate covariates at gldr.mean locations, using quadrilinear (3D + T) interpolation
}
