% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate4d.R, R/covariate4dOld.R
\name{covariate.4d}
\alias{covariate.4d}
\title{Calculate covariates at train.data locations and train.times,
using quadrilinear (3D + T) interpolation,
for a sigma coordinate system, where depth = f(lon, lat, depth, time)}
\usage{
covariate.4d(train.length, train.data, COV, train.times)

covariate.4d(train.length, train.data, COV, train.times)
}
\arguments{
\item{COV}{MASS-4D CHL}
}
\value{
Covariates at train.data locations and train.times.

Covariates at train.data locations and train.times.
}
\description{
Calculate covariates at train.data locations and train.times,
using quadrilinear (3D + T) interpolation,
for a sigma coordinate system, where depth = f(lon, lat, depth, time)

Calculate covariates at train.data locations and train.times,
using quadrilinear (3D + T) interpolation,
for a sigma coordinate system, where depth = f(lon, lat, depth, time)
}
