% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTime.R
\name{find.time}
\alias{find.time}
\title{Find nearest and neighbour time indices}
\usage{
find.time(dat.time.simulating, i, dat.time)
}
\arguments{
\item{dat.time.simulating}{dat.time.training, dat.time.predicting}

\item{dat.time}{mld.time, chl.time, ...}
}
\value{
Nearest and neighbour time indices.
}
\description{
Find nearest and neighbour time indices
}
