% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTimeIndex.R
\name{set.time.index}
\alias{set.time.index}
\title{Calculate training and predicting time indices}
\usage{
set.time.index(gldr.time.start, dat.time.start, cov.time, train.length, nt)
}
\arguments{
\item{dat.time.start}{time CMEMS & ECMWF records begin}

\item{cov.time}{covariate time values}
}
\value{
Training and predicting time indices.
}
\description{
Calculate training and predicting time indices
}
