% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTime.R
\name{set.time}
\alias{set.time}
\title{Calculate training and predicting times}
\usage{
set.time(gldr.time.start, cmems.time.start, train.length, nt)
}
\arguments{
\item{cmems.time.start}{time CMEMS records begin}
}
\value{
Training and predicting times.
}
\description{
Calculate training and predicting times
}
